+++ src/main/java/net/atlas/combatify/mixin/CombatRulesMixin.java
@@ -9,11 +9,13 @@ import net.atlas.combatify.config.ItemConfig;
 import net.atlas.combatify.enchantment.CustomEnchantmentHelper;
 import net.atlas.combatify.extensions.ItemExtensions;
 import net.atlas.combatify.extensions.LivingEntityExtensions;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.damagesource.CombatRules;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -24,9 +26,11 @@ public class CombatRulesMixin {
 	@Inject(method = "getDamageAfterAbsorb", at = @At("HEAD"))
 	private static void changeArmourAndProtection(LivingEntity attacked, float f, DamageSource damageSource, float g, float h, CallbackInfoReturnable<Float> cir, @Local(ordinal = 1, argsOnly = true) LocalFloatRef armour, @Local(ordinal = 2, argsOnly = true) LocalFloatRef toughness) {
 		if (Combatify.CONFIG.armourPiercingMode() == ArmourPiercingMode.APPLY_BEFORE_PERCENTAGE && damageSource.getEntity() instanceof LivingEntity livingEntity) {
-			Item item = livingEntity.getItemInHand(InteractionHand.MAIN_HAND).getItem();
+			ItemStack held = livingEntity.getItemInHand(InteractionHand.MAIN_HAND);
+			Item item = held.getItem();
 			double d = ((ItemExtensions)item).getPiercingLevel();
-			d += CustomEnchantmentHelper.getBreach(livingEntity);
+			if (!(attacked.level() instanceof ServerLevel serverLevel)) d += CustomEnchantmentHelper.getBreach(held, livingEntity.getRandom());
+			else d += CustomEnchantmentHelper.getArmorModifier(serverLevel, held, attacked, damageSource);
 			d -= ((LivingEntityExtensions)livingEntity).getPiercingNegation();
 			((LivingEntityExtensions)livingEntity).setPiercingNegation(0);
 			armour.set((float) (armour.get() * (1 - d)));